requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pymilvus==2.3.4

# Async & Scheduling
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# OpenAI
openai==1.3.7

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Finance Data
finance-datareader==0.9.50

# Telegram
python-telegram-bot==20.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
jinja2==3.1.2

# Authentication
bcrypt==4.1.3
passlib==1.7.4
python-jose[cryptography]==3.3.0
